# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "校园活动管理系统 API",
        "description": "校园活动管理系统接口文档",
        "contact": {
            "name": "开发团队",
            "email": "example@xidian.edu.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
        "/activity/update": {
            "put": {
                "tags": [
                    "活动管理"
                ],
                "summary": "更新活动",
                "description": "更新活动信息",
                "operationId": "updateActivity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/status": {
            "put": {
                "tags": [
                    "活动管理"
                ],
                "summary": "更新活动状态",
                "description": "修改活动状态",
                "operationId": "updateActivityStatus",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityStatusUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/publish/{activityId}": {
            "post": {
                "tags": [
                    "活动管理"
                ],
                "summary": "发布活动",
                "description": "发布活动，状态改为报名中",
                "operationId": "publishActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/create": {
            "post": {
                "tags": [
                    "活动管理"
                ],
                "summary": "创建活动",
                "description": "管理员创建新的活动",
                "operationId": "createActivity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityCreateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/cancel/{activityId}": {
            "post": {
                "tags": [
                    "活动管理"
                ],
                "summary": "取消活动",
                "description": "取消活动，状态改为已取消",
                "operationId": "cancelActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/{activityId}": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "获取活动详情",
                "description": "根据活动ID获取详细信息",
                "operationId": "getActivityDetail",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "活动管理"
                ],
                "summary": "删除活动",
                "description": "删除指定活动",
                "operationId": "deleteActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/list": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "查询活动列表",
                "description": "分页查询活动列表，支持筛选和排序",
                "operationId": "getActivityList",
                "parameters": [
                    {
                        "name": "queryDTO",
                        "in": "query",
                        "description": "查询条件",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ActivityQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageInfoActivityListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ResultVoid": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "ActivityUpdateDTO": {
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityName": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动名称",
                        "example": "新生计算机基础讲座（更新版）"
                    },
                    "activityDescription": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动描述",
                        "example": "由计算机学院主办，面向全校新生的计算机入门讲座，新增了实践环节"
                    },
                    "activityType": {
                        "type": "string",
                        "description": "活动类型编码",
                        "example": "ACADEMIC"
                    },
                    "startTime": {
                        "type": "string",
                        "description": "活动开始时间",
                        "format": "date-time"
                    },
                    "endTime": {
                        "type": "string",
                        "description": "活动结束时间",
                        "format": "date-time"
                    },
                    "location": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动地点",
                        "example": "学术报告厅A201"
                    },
                    "organizer": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string",
                        "description": "主办方",
                        "example": "计算机学院"
                    },
                    "contactPerson": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string",
                        "description": "负责人姓名",
                        "example": "王老师"
                    },
                    "contactPhone": {
                        "maxLength": 20,
                        "minLength": 0,
                        "type": "string",
                        "description": "联系电话",
                        "example": "13812345678"
                    },
                    "maxParticipants": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "最大参与人数，null表示不限制",
                        "format": "int32",
                        "example": 200
                    },
                    "posterUrl": {
                        "type": "string",
                        "description": "活动海报URL",
                        "example": "https://example.com/poster.jpg"
                    }
                },
                "description": "活动更新信息"
            },
...
            "ResultActivityDetailVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActivityDetailVO"
                    }
                }
            },
            "ActivityStatusUpdateDTO": {
                "required": [
                    "activityId",
                    "activityStatus"
                ],
                "type": "object",
                "properties": {
                    "activityId": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityStatus": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "description": "活动状态：0-未发布,1-报名中,2-报名结束,3-进行中,4-已结束,5-已取消",
                        "format": "int32",
                        "example": 1
                    }
                },
                "description": "活动状态更新信息"
            },
...
            "ActivityCreateDTO": {
                "required": [
                    "activityName",
                    "activityType",
                    "contactPerson",
                    "contactPhone",
                    "endTime",
                    "location",
                    "organizer",
                    "registrationEndTime",
                    "registrationStartTime",
                    "startTime"
                ],
                "type": "object",
                "properties": {
                    "activityName": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动名称",
                        "example": "新生计算机基础讲座"
                    },
                    "activityDescription": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动描述",
                        "example": "由计算机学院主办，面向全校新生的计算机入门讲座"
                    },
                    "activityType": {
                        "type": "string",
                        "description": "活动类型编码",
                        "example": "ACADEMIC"
                    },
                    "startTime": {
                        "type": "string",
                        "description": "活动开始时间",
                        "format": "date-time"
                    },
                    "endTime": {
                        "type": "string",
                        "description": "活动结束时间",
                        "format": "date-time"
                    },
                    "location": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string",
                        "description": "活动地点",
                        "example": "学术报告厅A201"
                    },
                    "organizer": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string",
                        "description": "主办方",
                        "example": "计算机学院"
                    },
                    "contactPerson": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string",
                        "description": "负责人姓名",
                        "example": "王老师"
                    },
                    "contactPhone": {
                        "maxLength": 20,
                        "minLength": 0,
                        "type": "string",
                        "description": "联系电话",
                        "example": "13812345678"
                    },
                    "maxParticipants": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "最大参与人数，null表示不限制",
                        "format": "int32",
                        "example": 200
                    },
                    "registrationStartTime": {
                        "type": "string",
                        "description": "报名开始时间",
                        "format": "date-time"
                    },
                    "registrationEndTime": {
                        "type": "string",
                        "description": "报名截止时间",
                        "format": "date-time"
                    },
                    "posterUrl": {
                        "type": "string",
                        "description": "活动海报URL",
                        "example": "https://example.com/poster.jpg"
                    }
                },
                "description": "活动创建信息"
            },
...
            "ActivityQueryDTO": {
                "type": "object",
                "properties": {
                    "pageNum": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "format": "int32",
                        "example": 1,
                        "default": 1
                    },
                    "pageSize": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页大小",
                        "format": "int32",
                        "example": 10,
                        "default": 10
                    },
                    "keyword": {
                        "type": "string",
                        "description": "活动名称关键词",
                        "example": "计算机"
                    },
                    "activityType": {
                        "type": "string",
                        "description": "活动类型编码",
                        "example": "ACADEMIC"
                    },
                    "activityStatus": {
                        "type": "integer",
                        "description": "活动状态：0-未发布,1-报名中,2-报名结束,3-进行中,4-已结束,5-已取消",
                        "format": "int32",
                        "example": 1
                    },
                    "creatorId": {
                        "type": "integer",
                        "description": "创建者ID",
                        "format": "int64",
                        "example": 1
                    },
                    "sortBy": {
                        "type": "string",
                        "description": "排序字段：createTime,startTime,registrationCount",
                        "example": "createTime"
                    },
                    "sortOrder": {
                        "type": "string",
                        "description": "排序方向：ASC,DESC",
                        "example": "DESC"
                    }
                },
                "description": "活动查询请求"
            },
...
            "PageInfoActivityListVO": {
                "type": "object",
                "properties": {
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "list": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityListVO"
                        }
                    },
                    "pageNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startRow": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "endRow": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "prePage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nextPage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isFirstPage": {
                        "type": "boolean"
                    },
                    "isLastPage": {
                        "type": "boolean"
                    },
                    "hasPreviousPage": {
                        "type": "boolean"
                    },
                    "hasNextPage": {
                        "type": "boolean"
                    },
                    "navigatePages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "navigatepageNums": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "navigateFirstPage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "navigateLastPage": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ResultPageInfoActivityListVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageInfoActivityListVO"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "请输入JWT Token，格式：直接输入token即可（不需要Bearer前缀）",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}